implemented SCM-Manager subversion source
