+++ src/main/java/com/cloudogu/scmmanager/scm/ScmManagerSource.java
@@ -1,30 +1,19 @@
 package com.cloudogu.scmmanager.scm;
 
-import com.cloudbees.plugins.credentials.common.StandardUsernameCredentials;
-import com.cloudbees.plugins.credentials.common.StandardUsernameListBoxModel;
-import com.cloudbees.plugins.credentials.domains.URIRequirementBuilder;
 import com.cloudogu.scmmanager.HttpAuthentication;
-import com.cloudogu.scmmanager.scm.api.ApiClient;
 import com.cloudogu.scmmanager.scm.api.Authentications;
-import com.cloudogu.scmmanager.scm.api.Repository;
 import com.cloudogu.scmmanager.scm.api.ScmManagerApi;
 import com.cloudogu.scmmanager.scm.api.ScmManagerHead;
 import com.cloudogu.scmmanager.scm.api.ScmManagerObservable;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Strings;
-import de.otto.edison.hal.HalRepresentation;
+import com.google.common.base.Joiner;
 import edu.umd.cs.findbugs.annotations.CheckForNull;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import hudson.Extension;
 import hudson.Util;
 import hudson.model.Action;
-import hudson.model.Item;
-import hudson.model.Queue;
 import hudson.model.TaskListener;
 import hudson.scm.SCM;
-import hudson.security.ACL;
-import hudson.util.FormValidation;
-import hudson.util.ListBoxModel;
 import jenkins.scm.api.SCMEvent;
 import jenkins.scm.api.SCMHead;
 import jenkins.scm.api.SCMHeadCategory;
@@ -34,7 +23,6 @@ import jenkins.scm.api.SCMProbe;
 import jenkins.scm.api.SCMRevision;
 import jenkins.scm.api.SCMSource;
 import jenkins.scm.api.SCMSourceCriteria;
-import jenkins.scm.api.SCMSourceDescriptor;
 import jenkins.scm.api.SCMSourceEvent;
 import jenkins.scm.api.SCMSourceOwner;
 import jenkins.scm.api.trait.SCMSourceRequest;
@@ -47,31 +35,22 @@ import jenkins.scm.impl.form.NamedArrayList;
 import jenkins.scm.impl.trait.Discovery;
 import jenkins.scm.impl.trait.Selection;
 import jenkins.util.NonLocalizable;
-import org.acegisecurity.Authentication;
 import org.jenkinsci.Symbol;
-import org.kohsuke.stapler.AncestorInPath;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
-import org.kohsuke.stapler.QueryParameter;
 
 import javax.annotation.Nonnull;
 import java.io.IOException;
 import java.io.PrintStream;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
 import java.util.function.BiFunction;
 import java.util.function.Function;
-import java.util.function.Predicate;
-
-import static java.util.Collections.emptyList;
+import java.util.stream.Collectors;
 
 public class ScmManagerSource extends SCMSource {
 
@@ -91,7 +70,7 @@ public class ScmManagerSource extends SCMSource {
 
   @DataBoundConstructor
   public ScmManagerSource(String serverUrl, String repository, String credentialsId) {
-    this(serverUrl, repository, credentialsId, ScmManagerSource::createHttpClient, Authentications::new);
+    this(serverUrl, repository, credentialsId, ScmManagerApi::create, Authentications::new);
   }
 
   ScmManagerSource(
@@ -218,7 +197,7 @@ public class ScmManagerSource extends SCMSource {
 
   private BiFunction<String, HttpAuthentication, ScmManagerApi> getApiFactory() {
     if (apiFactory == null) {
-      apiFactory = ScmManagerSource::createHttpClient;
+      apiFactory = ScmManagerApi::create;
     }
     return apiFactory;
   }
@@ -228,10 +207,6 @@ public class ScmManagerSource extends SCMSource {
       authenticationsProvider = Authentications::new;
     }
     return authenticationsProvider;
-  }
-
-  private static ScmManagerApi createHttpClient(String value, HttpAuthentication authentication) {
-    return new ScmManagerApi(new ApiClient(value, authentication));
   }
 
   static {
@@ -288,126 +263,21 @@ public class ScmManagerSource extends SCMSource {
 
   @Extension
   @Symbol("scm-manager")
-  public static class DescriptorImpl extends SCMSourceDescriptor {
-
-    private final BiFunction<String, HttpAuthentication, ScmManagerApi> apiFactory;
-
-    private final Predicate<Repository> repositoryPredicate;
+  public static class DescriptorImpl extends ScmManagerSourceDescriptor {
 
     public DescriptorImpl() {
-      this(ScmManagerSource::createHttpClient, SCMBuilderProvider::isSupported);
-    }
-
-    public DescriptorImpl(BiFunction<String, HttpAuthentication, ScmManagerApi> apiFactory, Predicate<Repository> repositoryPredicate) {
-      this.apiFactory = apiFactory;
-      this.repositoryPredicate = repositoryPredicate;
+      super(ScmManagerApi::create, SCMBuilderProvider::isSupported);
     }
 
     @Nonnull
     @Override
     public String getDisplayName() {
-      return "SCM-Manager";
-    }
-
-    @SuppressWarnings("unused") // used By stapler
-    public FormValidation doCheckServerUrl(@QueryParameter String value) throws InterruptedException, ExecutionException {
-      String trimmedValue = value.trim();
-      if (Strings.isNullOrEmpty(trimmedValue)) {
-        return FormValidation.error("server url is required");
-      }
-      try {
-        URI uri = new URI(value);
-        if (!uri.isAbsolute()) {
-          return FormValidation.error("illegal URL format");
-        }
-        if (!uri.getScheme().startsWith("http")) {
-          return FormValidation.error("Only http or https urls accepted");
-        }
-      } catch (URISyntaxException e) {
-        return FormValidation.error("illegal URL format");
-      }
-
-
-      ScmManagerApi api = apiFactory.apply(value, x -> {
-      });
-      CompletableFuture<HalRepresentation> future = api.index();
-      return future
-        .thenApply(index -> {
-          if (index.getLinks().getLinkBy("login").isPresent()) {
-            return FormValidation.ok();
-          }
-          return FormValidation.error("api has no login link");
-        })
-        .exceptionally(e -> FormValidation.error(e.getMessage()))
-        .get();
-    }
-
-    @SuppressWarnings("unused") // used By stapler
-    public FormValidation doCheckCredentialsId(@AncestorInPath SCMSourceOwner context, @QueryParameter String serverUrl, @QueryParameter String value) throws InterruptedException, ExecutionException {
-      return validateCredentialsId(context, serverUrl, value, Authentications::new);
-    }
-
-    @VisibleForTesting
-    FormValidation validateCredentialsId(@AncestorInPath SCMSourceOwner context, @QueryParameter String serverUrl, @QueryParameter String value, Function<SCMSourceOwner, Authentications> authenticationsProvider) throws InterruptedException, ExecutionException {
-      if (doCheckServerUrl(serverUrl).kind != FormValidation.Kind.OK) {
-        return FormValidation.error("server url is required");
-      }
-      if (Strings.isNullOrEmpty(value)) {
-        return FormValidation.error("credentials are required");
-      }
-      Authentications authentications = authenticationsProvider.apply(context);
-      ScmManagerApi client = apiFactory.apply(serverUrl, authentications.from(serverUrl, value));
-      CompletableFuture<HalRepresentation> future = client.index();
-      return future
-        .thenApply(index -> {
-          if (index.getLinks().getLinkBy("me").isPresent()) {
-            return FormValidation.ok();
-          }
-          return FormValidation.error("login failed");
-        })
-        .exceptionally(e -> FormValidation.error(e.getMessage()))
-        .get();
-    }
-
-    @SuppressWarnings("unused") // used By stapler
-    public ListBoxModel doFillCredentialsIdItems(@AncestorInPath SCMSourceOwner context, @QueryParameter String value) {
-      if (context == null || !context.hasPermission(Item.CONFIGURE)) {
-        return new StandardUsernameListBoxModel().includeCurrentValue(value);
-      }
-      Authentication authentication = context instanceof Queue.Task
-        ? ((Queue.Task) context).getDefaultAuthentication()
-        : ACL.SYSTEM;
-      return new StandardUsernameListBoxModel()
-        .includeEmptyValue()
-        .includeAs(authentication, context, StandardUsernameCredentials.class, URIRequirementBuilder.fromUri(value).build())
-        .includeCurrentValue(value);
-    }
-
-    @SuppressWarnings("unused") // used By stapler
-    public ListBoxModel doFillRepositoryItems(@AncestorInPath SCMSourceOwner context, @QueryParameter String serverUrl, @QueryParameter String credentialsId, @QueryParameter String value) throws InterruptedException, ExecutionException {
-      return fillRepositoryItems(context, serverUrl, credentialsId, value, Authentications::new);
-    }
-
-    public ListBoxModel fillRepositoryItems(@AncestorInPath SCMSourceOwner context, @QueryParameter String serverUrl, @QueryParameter String credentialsId, @QueryParameter String value, Function<SCMSourceOwner, Authentications> authenticationsProvider) throws InterruptedException, ExecutionException {
-      ListBoxModel model = new ListBoxModel();
-      if (Strings.isNullOrEmpty(serverUrl) || Strings.isNullOrEmpty(credentialsId)) {
-        if (!Strings.isNullOrEmpty(value)) {
-          model.add(value);
-        }
-        return model;
-      }
-
-      Authentications authentications = authenticationsProvider.apply(context);
-      ScmManagerApi api = apiFactory.apply(serverUrl, authentications.from(serverUrl, credentialsId));
-      List<Repository> repositories = api.getRepositories().exceptionally(e -> emptyList()).get();
-      for (Repository repository : repositories) {
-        if (repositoryPredicate.test(repository)) {
-          String displayName = String.format("%s/%s (%s)", repository.getNamespace(), repository.getName(), repository.getType());
-          String v = String.format("%s/%s/%s", repository.getNamespace(), repository.getName(), repository.getType());
-          model.add(displayName, v);
-        }
-      }
-      return model;
+      List<String> typeList = SCMBuilderProvider.all()
+        .stream()
+        .map(SCMBuilderProvider::getType)
+        .collect(Collectors.toList());
+      String types = Joiner.on(", ").join(typeList);
+      return String.format("SCM-Manager (%s)", types);
     }
 
     @SuppressWarnings("unused") // used By stapler
@@ -458,15 +328,6 @@ public class ScmManagerSource extends SCMSource {
         new ChangeRequestSCMHeadCategory(new NonLocalizable("Pull Requests")),
         TagSCMHeadCategory.DEFAULT
       };
-    }
-
-    static {
-      Icons.register("icon-scm-manager-source");
-    }
-
-    @Override
-    public String getIconClassName() {
-      return "icon-scm-manager-source";
     }
 
   }
